New support to upload a directory of JSON schema file(s) via "opa eval --schema". Directory can contain schema file(s) for policy input document(s), and schema file(s) for contextual data document(s). These schema files are used then to improve static type checking and to get more precise error reports as you develop Rego code
Also, there is new support for adding annotations on Rules to specify the schemas to be used specifically for type checking the expressions within the scope of that Rule. It helps address issues with schema overloading, and provides even more precise type error reports for a Rego developer.

If your directory contains an default-input-schema.json file for the input's schema, that will become the default input schema for all rules. This can be overriden with other input schema files, as well and explicitly linked in the rule annotation.

For example: https://github.com/aavarghese/opa-schema-examples/blob/ruleAnnotation/acl/

Documentation for new "rule annotation" feature in https://github.com/aavarghese/opa/blob/ruleAnnotation/docs/content/schemas.md

This work is continuation of merged PR #3060.

Co-authored-by: @vazirim Mandana Vaziri mvaziri@us.ibm.com
Co-authored-by: @aavarghese Ansu Varghese avarghese@us.ibm.com
Signed-off-by: Mandana Vaziri <mvaziri@us.ibm.com>